requirements.txt
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
Django==2.2.6
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
py==1.8.1
pylint==2.5.3
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
toml==0.10.1
urllib3==1.25.6
wcwidth==0.1.8
wrapt==1.12.1
zipp==2.2.0
